UEFI: parse ELF header
