+++ xv7-bootloader-uefi/src/main.rs
@@ -5,6 +5,8 @@
 #![feature(box_patterns)]
 #![feature(box_syntax)]
 
+mod io;
+
 #[macro_use]
 extern crate alloc;
 #[macro_use]
@@ -13,9 +15,6 @@ extern crate log;
 use alloc::vec::Vec;
 use chrono::prelude::*;
 use uefi::prelude::*;
-use uefi::proto::media::file::*;
-use uefi::proto::media::fs::SimpleFileSystem;
-use uefi::Result;
 
 #[no_mangle]
 extern "C" fn __rust_probestack() {}
@@ -53,14 +52,12 @@ fn efi_main(_image: Handle, system_table: SystemTable<Boot>) -> Status {
     );
 
     info!(r"Load kernel image from \EFI\xv7\kernel");
-    let (len, data) =
-        read_file(boot_services, r"\EFI\xv7\kernel").expect_success("Could not load kernel image");
+    let (len, data) = io::read_file(boot_services, r"\EFI\xv7\kernel")
+        .expect_success("Could not load kernel image");
 
     info!("Kernel image size = {}", len);
-    info!(
-        "Check kernel image magic number: {}{}{}",
-        data[1] as char, data[2] as char, data[3] as char
-    );
+
+    deal_with_elf(data).expect("ELF processing failed");
 
     loop {
         unsafe {
@@ -69,42 +66,9 @@ fn efi_main(_image: Handle, system_table: SystemTable<Boot>) -> Status {
     }
 }
 
-pub fn read_file(service: &BootServices, path: &str) -> Result<(usize, Vec<u8>)> {
-    let fatfs = service
-        .locate_protocol::<SimpleFileSystem>()
-        .log_warning()?;
-    let fatfs = unsafe { &mut *fatfs.get() };
-
-    let mut volume = fatfs.open_volume().log_warning()?;
-
-    let file_handle = volume
-        .open(path, FileMode::Read, FileAttribute::empty())
-        .log_warning()?;
-
-    let mut file = match file_handle.into_type().log_warning()? {
-        FileType::Regular(regular) => regular,
-        FileType::Dir(_) => unreachable!(),
-    };
-
-    // Use an empty buffer to retrieve the actual FileInfo size
-    let mut empty_buf = Vec::new();
-    let len = match *file
-        .get_info::<FileInfo>(&mut empty_buf)
-        .expect_error("passing an empty buffer will return the size of FileInfo")
-        .data()
-    {
-        Some(len) => len,
-        None => unreachable!(),
-    };
-
-    let mut file_info = vec![0u8; len];
-    let file_info = file
-        .get_info::<FileInfo>(&mut file_info)
-        .discard_errdata()
-        .log_warning()?;
-
-    let mut buf = vec![0u8; file_info.file_size() as usize];
-    let len = file.read(&mut buf).discard_errdata().log_warning()?;
-
-    Ok((len, buf).into())
+fn deal_with_elf(raw: Vec<u8>) -> core::result::Result<(), &'static str> {
+    let elf = xmas_elf::ElfFile::new(&raw)?;
+    info!("Kernel image info:");
+    info!("{}", elf.header);
+    Ok(())
 }
